fixing clippy errors

+++ tracewasm-core/src/runtime/stack.rs
@@ -424,7 +424,7 @@ impl RuntimeFrame for Stack<Value> {
     /// would strand this frame's locals on the stack under the caller's next push,
     /// which is the confusion the two names exist to prevent.
     fn exit_frame(&mut self, results_count: u32, caller_base_data: &Self::CallerBaseData) {
-        self.truncate_by_preserving_arity(caller_base_data.base_height, results_count as u32);
+        self.truncate_by_preserving_arity(caller_base_data.base_height, results_count);
     }
 }
